+++ source/extensions/omni.flux.stage_manager.plugin.listener.usd/omni/flux/stage_manager/plugin/listener/usd/stage_listener.py
@@ -37,7 +37,7 @@ class StageManagerUSDStageListenerPlugin(_StageManagerUSDListenerPlugin[omni.usd
         self._stage_event_sub = None
 
     def setup(self):
-        context = omni.usd.get_context(self.context_name)
+        context = omni.usd.get_context(self._context_name)
         self._stage_event_sub = context.get_stage_event_stream().create_subscription_to_pop(
             self._on_stage_event, name="StageManagerStageEventListener"
         )