+++ source/extensions/lightspeed.trex.stage_manager.plugin.widget.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.0]
+### Changed
+- Use set_context_name()
+
 ## [1.2.0]
 ### Changed
 - Use updated item data structure. `item.data` is now the prim itself.